+++ packages/rune-games-cli/src/flows/Upload/ChooseGameStep.tsx
@@ -2,8 +2,10 @@ import React, { useState, useMemo, useCallback, useEffect } from "react"
 
 import { Select } from "../../components/Select.js"
 import { Step } from "../../components/Step.js"
-import { useGames, useMyGames, gameItemLabel } from "../../gql/useGames.js"
+import { useGames } from "../../gql/useGames.js"
 import { useMe } from "../../gql/useMe.js"
+import { gameItemLabel } from "../../lib/gameItemLabel.js"
+import { useMyGames } from "../../lib/useMyGames.js"
 
 export function ChooseGameStep({
   currentGameId,
